Test: cts: RemoteStonithd test is not applicable if pacemaker service is enabled at boot
Cluster nodes would be reused as remote nodes in pacemaker-remote tests.
For RemoteStonithd, if pacemaker service was enabled at boot, after the
remote node got fenced, the cluster node would join instead of the
expected remote one. Meanwhile pacemaker_remote would not be able to
start. Depending on the chances, the situations might not be able to be
orchestrated gracefully any more.