Fix: controld: ensure existing node attributes get written back into CIB after daemons failed and recovered
If a pacemaker daemon unexpectedly exits, the daemon and probably also
other daemons including controld will try to get respawned and
recovered. Previously depending on which daemon had failed, if controld
got respawned and recovered either itself or together with some other
daemons, existing attributes of this node might not get written back
into CIB again.

It was found by the cts test "Reattach" randomly failing due to loss of
the node attribute "connected" in the CIB right after the test
"ComponentFail", for example if controld or execd was the failed
component.