Fix: scheduler: calculate digests without considering start-specific parameters
It's arguable that changes of op-specific parameters for non-recurring
operations should be taken into account when considering if resources
should be reloaded/restarted.

The fact in case of resulting into reload is, reload operations don't
really take any start-specific parameters, but a successful reload
gets updated as a start in cib resource history based on
pcmk__create_history_xml(), which of course will still keep the
same digests.