Fix: scheduler: reload takes any extra parameters configured for start op
For a reload action, if no reload op is configured or the reload op
doesn't have its own extra parameters, take start op's if there's
any.

A fact is, changes of start op's extra parameters result into a
reload, and a successful reload gets updated as a start lrm_rsc_op
based on pcmk__create_history_xml(). So this also ensures the digests
calculated for the resulting start lrm_rsc_op will be correct.

So far, not sure how to handle digests well if there's reload op
configured with extra parameters, although it's not an usual case.