Fix: scheduler: calculate secure digest based on all parameters only except sensitive ones
This reverts the behavior introduced by 6830621.

A secure digest is used for checking whether it's under the situation,
where among *all* the parameters, only any sensitive ones have changed.
Only if so, crm_simulate should ignore the changes.

For that purpose, a secure digest is supposed to be calculated based on
*all* the parameters only except the sensitive ones.