Fix: scheduler: restart a resource configured with extra operation parameters upon changes of any parameters
Given a non-recurring operation with extra parameters configured,
in case that the main digest doesn't match, even if the restart
digest matches, enforce a restart rather than a reload-agent anyway.
So that it ensures any changes of the extra parameters get applied
for this specific operation, and the digests calculated for the
resulting lrm_rsc_op will be correct.