Fix: libpacemaker: correctly find the best operation history entry to calculate digests for
Previously if there were multiple non-recurring operation entries, an
outdated operation might be picked. Especially given that migration
operations are now recorded separately with 48a97bcc1, the latest and
effective non-recurring operation should be preferred.

Not sure if it always makes sense, the logic for preferring non-failure
operations and operations with a restart digest is still preserved.