Fix: fencer: prevent watchdog fencing from timing out on insufficient `stonith-timeout`
A specified/defaulted global `stonith-timeout` shorter than
`stonith-watchdog-timeout` could be arguably considered a
misconfiguration. But given the configured `stonith-watchdog-timeout`,
it's pretty well known how long a watchdog fencing takes. In that case,
we should ideally tolerate the misconfiguration and allow watchdog
fencing to take as much time as it needs.

In any case, `stonith-timeout` still matters at least as the query
timeout.